designKit 오류 수정
