+++ Hedge/Projects/CoreKit/DesignKit/Sources/Components/HedgeNavigationBar.swift
@@ -29,7 +29,7 @@ public struct HedgeNavigationBar: View {
             Button {
                 onLeftButtonTap?()
             } label: {
-                Image(.chevronLeftThickSmall)
+                Image.hedgeUI.chevronLeftThickSmall
             }
             .frame(width: 40, height: 40)
